refactor(llmobs): export formatKnuthRate from priority_sampler.js
Keep the formatter in priority_sampler.js (export it) instead of
hoisting to sampler.js, per review.
