+++ packages/dd-trace/src/priority_sampler.js
@@ -17,7 +17,6 @@ const {
 const log = require('./log')
 const RateLimiter = require('./rate_limiter')
 const Sampler = require('./sampler')
-const { formatKnuthRate } = Sampler
 const { setSamplingRules } = require('./startup-log')
 const SamplingRule = require('./sampling_rule')
 
@@ -36,6 +35,19 @@ const {
 } = require('./constants')
 
 const DEFAULT_KEY = 'service:,env:'
+
+/**
+ * Formats a sampling rate as a string with up to 6 decimal digits and no trailing zeros.
+ *
+ * @param {number} rate
+ */
+function formatKnuthRate (rate) {
+  const string = Number(rate).toFixed(6)
+  for (let i = string.length - 1; i > 0; i--) {
+    if (string[i] === '0') continue
+    return string.slice(0, i + (string[i] === '.' ? 0 : 1))
+  }
+}
 
 const defaultSampler = new Sampler(AUTO_KEEP)
 
@@ -384,3 +396,4 @@ class PrioritySampler {
 }
 
 module.exports = PrioritySampler
+module.exports.formatKnuthRate = formatKnuthRate
